requirements.txt
pdfplumber
spacy
streamlit
huggingface-hub
python-dotenv
pandas